backend/requirements.txt
# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Database Drivers
motor>=3.3.2
pymongo>=4.6.1
qdrant-client>=1.7.3

# ML & Embeddings (Compatible with Python 3.14)
sentence-transformers>=2.3.1
torch>=2.9.0
torchvision>=0.20.0
pillow>=10.0.0
numpy>=1.26.0

# LLM Integration
google-genai>=0.2.0

# Data Processing
datasets>=2.16.1
pandas>=2.2.0
tqdm>=4.66.1

# Configuration & Validation
pydantic>=2.5.3
pydantic-settings>=2.1.0
python-dotenv>=1.0.1

# Utilities
httpx>=0.26.0
aiofiles>=23.2.1
